backend/requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.23.2
tensorflow-cpu==2.10.0
pillow==10.0.0
python-multipart==0.0.6
numpy==1.23.5
h5py==3.7.0
